requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
psycopg2-binary==2.9.9

openai==1.52.0   # latest stable OpenAI SDK

python-dotenv==1.0.1   # optional, for environment variables
